separate search.rs
